Issu of exc cmmd

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,4 +1,8 @@
+# Using Puppet, create a manifest that kills a process named killmenow
+
 exec { 'pkill':
   command  => 'pkill killmenow',
-  provider => 'shell',
+  onlyif  => '/usr/bin/pgrep killmenow',
+  refreshonly => true,
+  provider => 'shell'
 }
